adding video functionality, authentication changes-login with credentials, permissions denied dialog boxes

+++ components/DropDownMenuPost.js
@@ -7,20 +7,37 @@ import Router from 'next/router';
 import DialogBox from './DialogBox';
 import { useDispatch } from 'react-redux';
 import { deletePost } from '@/public/src/features/postSlice';
+import { useSession } from 'next-auth/react';
 
 function classNames(...classes) {
   return classes.filter(Boolean).join(' ')
 }
 
-export default function DropDownMenuPost({postId}) {
+export default function DropDownMenuPost({postId, authorEmail}) {
     const DELETE_API_ENDPOINT =`${process.env.NEXT_PUBLIC_PAGE_BASEURL}api/v1/posts/${postId}`;
     const [modalOpen, setModalOpen] = useState(false);
+    const [noPermissionModalOpen, setnoPermissionModalOpen] = useState(false);
+    const {data: session} = useSession();
     const handleSucces = () => {
       Router.push('/')
+    }
+    const handleEdit = () => {
+      if(session?.user.email!=authorEmail) {
+        setnoPermissionModalOpen(true);
+        return;
+      }
+      Router.push(`/posts/edit/${postId}`);
+    }
+    const handleNoPermission = () => {
+      setnoPermissionModalOpen(false);
     }
     const dispatch = useDispatch();
     const deletePostById = (e) => {
         e.preventDefault();
+        if(session?.user.email!=authorEmail) {
+          setnoPermissionModalOpen(true);
+          return;
+        }
         axios.delete(DELETE_API_ENDPOINT,{
             headers:{
               Accept:"application/json" 
@@ -56,16 +73,10 @@ export default function DropDownMenuPost({postId}) {
           <div className="py-1">
             <Menu.Item>
               {({ active }) => (
-                 <Link href={{
-                    pathname: `/posts/edit/[postId]`,
-                    query: { postId: postId},
-                    }} 
-                  className={classNames(
-                    active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
-                    'block px-4 py-2 text-sm'
-                  )}
-                > Edit
-                </Link>
+                <div onClick={handleEdit} className={classNames(
+                  active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                  'block px-4 py-2 text-sm cursor-pointer'
+                )}>Edit</div>
               )}
             </Menu.Item>
             <Menu.Item>
@@ -84,6 +95,7 @@ export default function DropDownMenuPost({postId}) {
         </Menu.Items>
       </Transition>
       {modalOpen &&(<DialogBox messageHead="Post deleted sucessfully!" message="Post was deleted sucessfully. You will now be redirected to homepage" handleSucces={handleSucces}/>)}
+      {noPermissionModalOpen &&(<DialogBox messageHead="You dont have permission to do that!" message="You are not the author of this post, so you can't delete or modify it." handleSucces={handleNoPermission}/>)}
     </Menu>
   )
 }
